♻️ commonコンポーネントがページコンポーネントに依存しないようにした (#61)
