+++ src/_components/common/BarsDataset/BarsDataset.tsx
@@ -2,8 +2,13 @@
 
 import { BarChart } from "@mui/x-charts/BarChart";
 import { axisClasses } from "@mui/x-charts/ChartsAxis";
-import { BarsDatasetType } from "@/_components/features/dashboard/type";
 import { FC } from "react";
+
+export type BarsDatasetType = {
+  budget: number;
+  expense: number;
+  categoryName: string;
+};
 
 type BarsDatasetProps = {
   dataset: BarsDatasetType[];
