--- src/_components/features/dashboard/type.ts
@@ -1,12 +1,0 @@
-export type PieChartData = {
-  id: number;
-  value: number;
-  label: string;
-  color: string;
-};
-
-export type BarsDatasetType = {
-  budget: number;
-  expense: number;
-  categoryName: string;
-};
